==> 01_extract_Aufgaben_dataset.ipynb <==
import pandas as pd 
df = pd.read_parquet("results.parquet")

def parse_text(text): 
    topics = text.split("\n\n")

    data = []

    for topic in topics: 
        if topic: 
            title, *questions = topic.split("\n")
            title = title.removesuffix("':").removeprefix("'")
            questions = [question.split(". ",1)[1].removesuffix('"').removeprefix('"') if ". " in question else question for question in questions]
            data.append({"topic": title, "questions": questions})
    return pd.DataFrame(data)

df_parsed = pd.concat(df["results"].apply(parse_text).tolist(), ignore_index=True)

num_questions = df_parsed["questions"].apply(len).sum()
print(f"Generated {num_questions} question topics")
df_parsed

df_parsed.drop_duplicates(subset=["topic"])#.sample(4)

print("Unique topic questions:", df_parsed.drop_duplicates(subset=["topic"])["questions"].apply(len).sum())

df_parsed.to_parquet("results_parsed.parquet")
