==> 03_parse_Questions.ipynb <==
import pandas as pd 

df = pd.read_parquet("results_questions.parquet")
df["gen_questions"][30]

samples = df.dropna(subset="gen_questions").iloc[:7*32].sample(3)
for sample in samples["gen_questions"]:
    print(sample)
    print("-----------------------------")
